% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_admission_status.R
\name{check_admission_status}
\alias{check_admission_status}
\title{Title : Did you get accepted to Hogwarts?}
\usage{
check_admission_status(name)
}
\arguments{
\item{name}{A character string the represents the name of a student.}
}
\value{
Returns a statement with name and admission decision.
}
\description{
This function generates the admission decision to the Hogwarts School of Witchcraft and Wizardry
}
\examples{

check_admission_status(name = "Mubarika")
}
