% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance.R
\name{letter}
\alias{letter}
\title{Title : Did you get accepted to Hogwarts?}
\usage{
letter(name)
}
\arguments{
\item{name}{A character string the represents the name of a student.}
}
\value{
Returns a statement with name and admission decision.
}
\description{
This function generates the admission decision to the Hogwarts School of Witchcraft and Wizardry
}
\examples{

letter(name = "Mubarika")
}
