% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_into_house.R
\name{sort_into_house}
\alias{sort_into_house}
\title{Title : Which Hogwarts house do you belong}
\usage{
sort_into_house(name)
}
\arguments{
\item{name}{A character string the represents the name of a student.}
}
\value{
Returns a statement with name and assigned house.
}
\description{
This function acts as the sorting hat and randomly assigns the
given name to a Hogwarts house.
}
\details{
This function acts as the sorting hat and randomly assigns the
given name to a Hogwarts house.
}
\examples{

sort_into_house(name = "Mubarika")
}
