% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_hat.R
\name{sorting_hat}
\alias{sorting_hat}
\title{Title : Which Hogwarts house do you belong}
\usage{
sorting_hat(name)
}
\arguments{
\item{name}{A character string the represents the name of a student.}
}
\value{
Returns a statement with name and assigned house.
}
\description{
This function acts as the sorting hat and randomly assigns the
given name to a Hogwarts house.
}
\details{
This function acts as the sorting hat and randomly assigns the
given name to a Hogwarts house.
}
\examples{

sorting_hat(name = "Mubarika")
}
